Update Image and Documentation
Update the image to the latest. Also update documentation
links to refer to this location for images built here.

New release with auto created assets
Automate releases using github actions to create the tarball.

Why do you need the tarball.

Inception: Add support to auto build and upload latest tarball.
This tarball is needed to solve the chicken and egg problem.
This allows us to use wget to grab the docker image through
a proxy.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>