Fix install on non-nft systems
Current version of the script looks for `iptables -L nat` on the system
to determine whether iptables or nftables are installed. In latest
Ubuntu 20.04 server atleast, nat tables do not exist and that command
will fail leading to the container assuming nftables is installed. This
is a small change to fix that scenario.

Signed-off-by: Ganesh Maharaj Mahalingam <ganesh.mahalingam@intel.com>